requirements.txt
streamlit
google-generativeai
pandas
matplotlib
python-dotenv
pillow